add file for render functions and fixed html, css, js
