+++ src/main.js
@@ -2,8 +2,10 @@ import axios from 'axios';
 import Swiper from 'swiper';
 import 'swiper/css';
 import { Notify } from 'notiflix';
-import SlimSelect from 'slim-select'
+import SlimSelect from 'slim-select';
 import debounce from 'lodash.debounce';
 
 import './js/hero.js';
 import './js/fillCategories';
+import './js/popular-recipes.js';
+import './js/render-popular.js';
