+++ src/partials/popular-recipes.html
@@ -1,8 +1,8 @@
-<section>
-  <div class="container">
-    <div class="popular-recipes-container">
-      <h2 class="popular-recipe-title">Popular recipes</h2>
-      <div class="popular-recipes-card"></div>
-    </div>
-  </div>
-</section>
+<!-- <section> -->
+<!-- <div class="container"> -->
+<div class="popular-recipes-container">
+  <h2 class="popular-recipe-title">Popular recipes</h2>
+  <div class="popular-recipes-card"></div>
+</div>
+<!-- </div> -->
+<!-- </section> -->
